requirements.txt
aiofiles==0.8.0
aioredis==2.0.1
aiosmtplib==1.1.6
aiosqlite==0.17.0
ansible==4.4.0
ansible-core==2.11.4
anyio==3.4.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.4.1
asn1crypto==1.4.0
async-timeout==4.0.2
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
Brlapi==0.7.0
CacheControl==0.12.10
cached-property==1.5.2
cachy==0.3.0
cbor2==5.4.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cleo==0.8.1
Click==7.0
clikit==0.6.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
crashtest==0.3.1
cryptography==3.4.7
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.4
distro==1.6.0
distro-info===0.23ubuntu1
Django==3.2.8
django-debug-toolbar==3.2.2
django-filter==21.1
djangorestframework==3.12.4
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
email-validator==1.1.3
entrypoints==0.3
fakeredis==1.7.0
fastapi==0.70.0
fastapi-mail==1.0.2
fasteners==0.16.3
filelock==3.4.2
future==0.18.2
h11==0.12.0
html5lib==1.1
httpcore==0.13.7
httplib2==0.19.1
httpx==0.19.0
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
iso8601==0.1.16
isodate==0.6.0
jeepney==0.7.1
Jinja2==3.0.1
jsonschema==3.2.0
keyring==21.8.0
language-selector==0.1
launchpadlib==1.10.14
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.9.1
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==4.2.0
msgpack==1.0.3
netifaces==0.11.0
oauthlib==3.1.1
olefile==0.46
packaging==20.9
paramiko==2.7.2
passlib==1.7.4
pastel==0.2.1
pbr==5.6.0
pexpect==4.8.0
Pillow==9.0.0
piwikapi==0.3
pkginfo==1.8.2
platformdirs==2.2.0
pluggy==1.0.0
poetry==1.1.12
poetry-core==1.0.7
protobuf==3.18.0
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pycairo==1.20.1
pycparser==2.20
pycups==2.0.1
pydantic==1.8.2
Pygments==2.10.0
PyGObject==3.42.0
PyJWT==2.1.0
pylev==1.4.0
pymacaroons==0.13.0
pymongo==4.0.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypika-tortoise==0.1.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyRFC3339==1.1
pyrsistent==0.18.0
pytest==6.2.5
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian==0.1.40
python-dotenv==0.19.1
python-multipart==0.0.5
pytz==2021.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.0.2
remarshal==0.14.0
reportlab==3.6.1
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
resolvelib==0.5.5
rfc3986==1.5.0
SecretStorage==3.3.1
shellingham==1.4.0
simplejson==3.17.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
starlette==0.16.0
systemd-python==234
testresources==2.0.1
texttable==1.6.4
toml==0.10.2
tomlkit==0.7.2
tortoise-orm==0.18.0
typing_extensions==4.0.1
u-msgpack-python==2.7.1
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.6
usb-creator==0.3.7
uvicorn==0.15.0
virtualenv==20.12.0
wadllib==1.3.6
webencodings==0.5.1
websocket-client==0.59.0
wrapt==1.13.3
xkit==0.0.0
zeep==4.1.0
zipp==3.5.0
